+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - **jsdoc** config: added a new option `extraMultilineCommentsStartingWithToIgnore` to control which multiline comments should be ignored by [`no-bad-blocks`](https://github.com/gajus/eslint-plugin-jsdoc/blob/HEAD/docs/rules/no-bad-blocks.md) rule. Added `__PURE__` and `__NO_SIDE_EFFECTS__` to the default list.
 - **casePolice** config: autofix is now disabled by default.
 - **react** config: `configTypeAwareRules` now uses `files` and `ignores` from `ts` config by default.
+- **yaml** config: add a new config to target GitHub Actions workflow files, currently only disabling [`no-empty-mapping-value`](https://ota-meshi.github.io/eslint-plugin-yml/rules/no-empty-mapping-value.html) rule.
 - Added a new root option, `defaultConfigsStatus`, to control what configs are enabled or disabled by default.
 
 ## 0.10.0